+++ website/management/commands/run_daily.py
@@ -45,6 +45,10 @@ class Command(BaseCommand):
                 call_command("daily_checkin_reminder")
             except Exception as e:
                 logger.error("Error sending daily checkin reminders", exc_info=True)
+            try:
+                call_command("cron_send_reminders")
+            except Exception as e:
+                logger.error("Error sending user reminders", exc_info=True)
         except Exception as e:
             logger.error("Error in daily tasks", exc_info=True)
             raise
